+++ src/main/java/Team_Mute/back_end/domain/reservation_admin/repository/ReservationLogRepository.java
@@ -1,9 +1,10 @@
 package Team_Mute.back_end.domain.reservation_admin.repository;
 
-import Team_Mute.back_end.domain.reservation_admin.entity.ReservationLog;
-
 import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import Team_Mute.back_end.domain.reservation_admin.entity.ReservationLog;
 
 /**
  * [예약 관리 -> 반려 사유 로그] 리포지토리
@@ -20,5 +21,5 @@ public interface ReservationLogRepository extends JpaRepository<ReservationLog, 
 	 * @return 가장 최근의 ReservationLog 기록 Optional
 	 */
 	Optional<ReservationLog> findTopByReservationReservationIdAndChangedStatusReservationStatusIdOrderByRegDateDesc(
-		Long reservationId, Long statusId);
+		Long reservationId, Integer statusId);
 }
